add class

+++ .eslintrc.js
@@ -30,7 +30,7 @@ module.exports = {
     'prettier/prettier': 'error',
     'react/prop-types': 0,
     'linebreak-style': ['error', 'unix'],
-    quotes: ['warn', 'single'],
+    quotes: ['warn', 'double'],
     semi: ['warn', 'always'],
     'comma-dangle': ['error', 'always-multiline'],
   },
